update the structure of the evaluators
